+++ paper.bib
@@ -2,6 +2,7 @@
   title={The Orthanc Ecosystem for Medical Imaging},
   author={Jodogne, S{\'e}bastien},
   journal={Journal of Digital Imaging},
+  doi = {10.1007/s10278-018-0082-y},
   volume={31},
   number={3},
   pages={341--352},
