### 2024-09-23 #7236 @gacou54
Thanks @Kevin-Mattheus-Moerman , `pyorthanc` doesn't have much domain logic. It's mainly a wrapper to communicate with Orthanc's REST API (Orthanc is an open-source medical images server) with utility functions. So I think anyone with Python + REST API knowledge could review it.

`pyorthanc` is already used by many, so (I hope :) ) the library has proven its relevance and deserves publication.
