### 2024-09-23 #7236 @Kevin-Mattheus-Moerman
@gacou54 thanks for your comments. Although I'm convinced this represents a significant piece of work, that is no doubt very useful, I am a little worried if this work is in scope for JOSS, e.g. our guidelines say `“Minor utility” packages, including “thin” API clients, and single-function packages are not acceptable`. Although this submission seems substantial enough in terms of number of lines of code (so not very thin), can it arguably be defined as a wrapper/single-function utility package?

JOSS publishes "research software", i.e. software that is used for scientific research. A good check for JOSS scope is to image if one would feel the need to cite this work when used in such scientific research. There is typically no need to cite a utility tool but one would cite something that would influence the scientific findings directly. So can you help sketch a scenario where people would use your software in scientific research and where people would feel the need to cite your package because it influences the scientific findings/conclusions? Has this work been used and cited in scientific research papers already?

Any clarification on the above would be very helpful to understand if this represents research software. Thanks.
