### 2024-09-25 #7236 @gacou54
I understand the issue you raised. PyOrthanc is used in several research projects and is an important element in medical imaging data management projects.

PyOrthanc is typically used upstream in anonymization and data transfer flows, and downstream for complex data management and queries. On our side, we are working on a medical imaging platform dedicated to AI and other algorithms in research (https://paradim.science/), and we orchestrate the data with PyOrthanc.

One of the reasons why we would like to publish in JOSS is that we often see that PyOrthanc is used at scientific conferences for medical imaging projects and we would like people to cite our work.

We will rely on your judgment on its publication value. While it is true that PyOrthanc is a fancy wrapper with several utility functions, its main purpose and the reason we created this tool is to support medical image research. In our opinion, PyOrthanc is a real asset for many medical imaging data management endeavors.
