### 2024-11-07 bff041d
✨Orthanc docker image with pyorthanc
