### 2024-11-20 #7236 @sappelhoff
@gacou54 do you have suggestions for potential reviewers? Please just have potential [conflicts of interest](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) in mind.

If not, I will consult the automatic suggestions [from above](https://github.com/openjournals/joss-reviews/issues/7236#issuecomment-2352846435), or will have a look at the [JOSS reviewer database](https://reviewers.joss.theoj.org/).
