### 2024-11-26 #7236 @gacou54
Hi @sappelhoff ,

I don't have anyone in particular in mind. I think anyone who has handled DICOM (medical imaging) files as part of their job might have a good idea of ​​what's going on.

The suggestions [here](https://github.com/openjournals/joss-reviews/issues/7236#issuecomment-2352846435) seem good to me :)
