### 2024-11-26 #7236 @sappelhoff
👋 @jrkerns @matteocao @sinhaharsh @drcandacemakedamoore @devnums

would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
