### 2024-12-02 #7236 @matteocao
Dear Stefan,

Thank you for reaching out.

Unfortunately this is a very intense working period and I would not have
the time to review the submission. Apologies.

Best regards,

Matteo

Il Lun 2 Dic 2024, 22:08 Stefan Appelhoff ***@***.***> ha
scritto:

> I have reached out to @devnums <https://github.com/devnums> @matteocao
> <https://github.com/matteocao> via email to see whether they would be
> interested to review this submission.
>
