### 2024-12-03 #7562 @editorialbot
[REVIEW]: PyOrthanc: A Python Interface for Orthanc DICOM Servers
**Submitting author:** @gacou54 (<a href="http://orcid.org/0000-0003-1832-6036">Gabriel Couture</a>)
**Repository:** https://github.com/gacou54/pyorthanc
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.18.0
