+++ docker/README.md
@@ -1,0 +1,5 @@
+#To build this dockerile :
+
+```shell
+docker build --build-arg PYORTHANC_VERSION=1.18.0 --build-arg ORTHANC_VERSION=24.9.1  .
+```
