### 2024-12-19 gacou54/pyorthanc#69 @devnums
The following line in the documentation doesn't work
This issue is related to [joss-review 7562](https://github.com/openjournals/joss-reviews/issues/7562)
The following line in the documentation doesn't work, because the demo server has been redirected to another address.

```client = pyorthanc.Orthanc('https://demo.orthanc-server.com/')```

Also, this part in the documentation doesn't work. Could you please clarify what should be the path?

```from pyorthanc import retrieve_and_write_patients

for patient in patients:
    patient.download(f'./data/patient-{patient.patient_id}.zip', with_progres=False)

# As a directory tree DICOM files (patients -> studies -> series -> instances)
retrieve_and_write_patients(patients, './data/')

```
