### 2024-12-19 gacou54/pyorthanc#70 @devnums
Please address these points in the documentation?
This issue is related to [joss-review 7562](https://github.com/openjournals/joss-reviews/issues/7562)

Is there a link to the  https://www.orthanc-server.com/, specially in the installation and usage sections?

Could you please mention the system configuration (os,...) that the software has been tested to work on?

Thank you
