### 2024-12-19 #7562 @devnums
Thank you @gacou54 for your useful software.

Could you please address the followings in the documentation.

In the documentation is there a statement of need?
Issues on the project repository: [#70](https://github.com/gacou54/pyorthanc/issues/70) and [#69](https://github.com/gacou54/pyorthanc/issues/69)
I also suggest adding _report issues_ and _seek supports_ sections in the documentations as well as more example usages in the documentation and the software repository covering major functionalities of the software.
