### 2024-12-19 gacou54/pyorthanc#69 @gacou54
Thanks for pointing that out! The URL of the demo server has indeed changed, I will fix this soon.
