### 2024-12-19 gacou54/pyorthanc#70 @gacou54
I will make this change soon.
