### 2024-12-19 1d7d497
Fix warnings when building the orthanc-pyorthanc dockerfile
