+++ docker/README.md
@@ -1,5 +1,5 @@
-#To build this dockerile :
+# To build this Dockerfile:
 
 ```shell
-docker build --build-arg PYORTHANC_VERSION=1.18.0 --build-arg ORTHANC_VERSION=24.9.1  .
+docker build --build-arg PYORTHANC_VERSION=1.18.0 --build-arg ORTHANC_VERSION=24.9.1  -t ylemarechal/orthanc-pyorthanc:24.9.1 .
 ```
