### 2024-12-19 cd32ddd
Fix warnings when building the orthanc-pyorthanc dockerfile
