### 2024-12-20 0168e48
Improve introduction documentation
