+++ docs/index.md
@@ -1,66 +1,123 @@
-<!-- ====================== TOC ====================== -->
-<!-- Generated by mkdocs-toc-md plugin -->
-<!-- ================================================= -->
-
 # PyOrthanc
-**PyOrthanc** is a python client for the Orthanc REST API, which fully wraps all the methods of the REST API.
-Additionally, it provides many utility functions to interact with an Orthanc instance.
 
+**PyOrthanc** is a Python client for the [Orthanc](https://www.orthanc-server.com/) REST API. It provides:
 
----
-Install PyOrthanc
+- Complete wrapping of the Orthanc REST API methods
+- High-level utilities for common DICOM operations
+- Asynchronous client support
+- Helper functions for working with DICOM data
+- Integration with the Orthanc Python plugin
+
+## Quick Install
+
 ```bash
-pip install pyorthanc
+pip install pyorthanc        # Basic installation
+pip install pyorthanc[all]   # Install all optional dependencies
 ```
----
-Then use the client:
-```python
-import pyorthanc
 
-client = pyorthanc.Orthanc('https://orthanc.uclouvain.be/demo')
+## Basic Usage
+```python
+from pyorthanc import Orthanc
+
+# Connect to Orthanc server
+client = Orthanc('http://localhost:8042')
+# Or with authentication:
+client = Orthanc('http://localhost:8042', username='orthanc', password='orthanc')
+
+# Basic operations
 patient_ids = client.get_patients()
+studies = client.get_studies() 
+
+# Upload DICOM file
+with open('image.dcm', 'rb') as f:
+    client.post_instances(f.read())
 ```
-Interact with connected modalities
+## Working with DICOM Modalities
 
 ```python
-import pyorthanc
+from pyorthanc import Modality
 
-modality = pyorthanc.Modality(client, 'MY_MODALITY')
-assert modality.echo()
+# Create modality connection
+modality = Modality(client, 'REMOTE_PACS')
 
-# C-Find on modality
-response = modality.query({'Level': 'Study', 'Query': {'PatientID': '*'}})
+# Test connection with C-ECHO
+if modality.echo():
+    print("Successfully connected to PACS")
 
-# C-Move to target modality
-modality.move(response['ID'], {'TargetAet': 'target_modality'})
+# Query studies with C-FIND
+response = modality.query({
+    'Level': 'Study',
+    'Query': {
+        'PatientID': '12345*',
+        'StudyDate': '20230101-20231231'
+    }
+})
+
+# Retrieve results with C-MOVE to a target AET
+modality.move(response['ID'], {'TargetAet': 'ORTHANC'})
 ```
-Find patients
+## Finding and Processing DICOM Data
+
 ```python
-patients = pyorthanc.find_patients(client, {'PatientName': '*Gabriel'})
+from pyorthanc import find_patients
+
+# Search for patients
+patients = find_patients(
+    client,
+    query={'PatientName': '*Gabriel'},
+    labels=['research']  # It is also possible to filter by labels
+)
+
+# Process patient data
 for patient in patients:
-    patient.labels
-    patient.is_stable
-    patient.name
-    ...
+    print(f"Patient: {patient.name} (ID: {patient.patient_id})")
+    print(f"Birth Date: {patient.birth_date}")
+    print(f"Labels: {patient.labels}")
+    
+    # Access studies
     for study in patient.studies:
-        study.labels
-        study.date
-        ...
+        print(f"\nStudy Date: {study.date}")
+        print(f"Description: {study.description}")
+        
+        # Access series
         for series in study.series:
-            ...
+            print(f"\nModality: {series.modality}")
[…]
