+++ README.md
@@ -2,80 +2,134 @@
 [![documentation](https://img.shields.io/badge/docs-mkdocs%20material-blue.svg?style=flat)](https://gacou54.github.io/pyorthanc/)
 ![PyPI - Version](https://img.shields.io/pypi/v/pyorthanc)
 ![PyPI - Downloads](https://img.shields.io/pypi/dm/pyorthanc)
+
 # PyOrthanc
-**PyOrthanc** is a python client for the Orthanc REST API, which fully wraps all the methods of the REST API.
-Additionally, it provides many utility functions to interact with an Orthanc instance.
 
-See the full documentation here https://gacou54.github.io/pyorthanc
+**PyOrthanc** is a comprehensive Python client for [Orthanc](https://www.orthanc-server.com/), providing:
 
----
-Install PyOrthanc using pip:
+- Complete wrapping of the Orthanc REST API methods
+- High-level utilities for common DICOM operations 
+- Asynchronous client support
+- Helper functions for working with DICOM data
+- Integration with the [Orthanc Python plugin](https://orthanc.uclouvain.be/book/plugins/python.html)
+
+## Quick Install
 ```bash
-pip install pyorthanc
+pip install pyorthanc        # Basic installation
+pip install pyorthanc[all]   # Install all optional dependencies
 ```
----
-Then use the client. If Orthanc is running locally, the default URL is `http://localhost:8042`.
-```python
-import pyorthanc
 
-client = pyorthanc.Orthanc('http://localhost:8042', username='orthanc', password='orthanc')
+## Basic Usage
+Assuming an Orthanc server running locally at `http://localhost:8042`:
+```python
+from pyorthanc import Orthanc
+
+# Connect to Orthanc server
+client = Orthanc('http://localhost:8042')
+# Or with authentication:
+client = Orthanc('http://localhost:8042', username='orthanc', password='orthanc')
+
+# Basic operations
 patient_ids = client.get_patients()
+studies = client.get_studies() 
+
+# Upload DICOM file
+with open('image.dcm', 'rb') as f:
+    client.post_instances(f.read())
 ```
 
-Interact with connected modalities
+## Working with DICOM Modalities
+
 ```python
-import pyorthanc
+from pyorthanc import Modality
 
-modality = pyorthanc.Modality(client, 'MY_MODALITY')
-assert modality.echo()
+# Create modality connection
+modality = Modality(client, 'REMOTE_PACS')
 
-# C-Find on modality
-response = modality.query({'Level': 'Study', 'Query': {'PatientID': '*'}})
+# Test connection with C-ECHO
+if modality.echo():
+    print("Successfully connected to PACS")
 
-# C-Move to target modality
-modality.move(response['ID'], {'TargetAet': 'target_modality'})
-```
-Find patients
-```python
-patients = pyorthanc.find_patients(client, {'PatientID': '*P001'})
-for patient in patients:
-    patient.labels
-    patient.is_stable
-    patient.name
-    ...
-    for study in patient.studies:
-        study.labels
-        study.date
-        ...
-        for series in study.series:
-            ...
-            for instance in series.instances:
-                pydicom_ds = instance.get_pydicom()
+# Query studies with C-FIND
+response = modality.query({
+    'Level': 'Study',
+    'Query': {
+        'PatientID': '12345*',
+        'StudyDate': '20230101-20231231'
+    }
+})
+
+# Retrieve results with C-MOVE to a target AET
+modality.move(response['ID'], {'TargetAet': 'ORTHANC'})
 ```
 
-Resources (`Patient`, `Study`, `Series`, `Instance`) can be easily __anonymized__.
-```python
-import pyorthanc
+## Finding and Processing DICOM Data
 
-orthanc_patient_id = client.get_patients()[0]
[…]
