### 2024-12-20 ad38b86
Add information on how to download/setup Orthanc locally for first steps
