### 2024-12-20 #7562 @gacou54
Thanks for the feedback!

> In the documentation is there a statement of need?

Only in the paper submitted in JOSS. However, as you suggested, I am improving a few parts of the documentation. The summary in the `README.md` will contain a message about the relevance of pyorthanc.

I am addressing the issues you raised regarding the documentation (https://github.com/gacou54/pyorthanc/pull/72). I also agree that there should be more examples. I will add them after the holidays. Thanks :)
