### 2025-01-11 gacou54/pyorthanc#73 @sinhaharsh
[JOSS] Trouble running tests on macOS
https://github.com/openjournals/joss-reviews/issues/7562

Operating System: macOS 15.2

## Description

I tried to run the tests as described in README.md
Please let me know where I should correct my approach.

## What I Did
docker compose run --build test

## Output
```
➜  docker compose run --build test
[+] Running 54/51
 ✔ orthanc2 Pulled                                                                                                                                                                                                                                                         0.8s
   ✔ 26070551e657 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 6cb16b7dc94f Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 94dad2d9840d Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 2217db394bdc Already exists                                                                                                                                                                                                                                           0.0s
   ✔ b51aa7af555a Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 86dbbef1723c Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 3d957f40cdaf Already exists                                                                                                                                                                                                                                           0.0s
   ✔ af721ba22ca9 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ c4af6f54addd Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 8746cf37628d Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 9a3ce58e58d0 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ c797a3ca1597 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 9e7ee13b7bc6 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 3ae15a01d2e4 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 730c81894416 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ e2cf1c3c98fe Already exists                                                                                                                                                                                                                                           0.0s
   ✔ a923a9e001d2 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 06583b10a86f Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 8b2c5e4ec063 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 435e71ffd689 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 6987982bc4bf Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 3cf5407f22b5 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 0572e029bc52 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ c953d2be32c5 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 1c898a651891 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ ea87c8842a3d Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 93d281d6e3ff Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 46d2aaa19e4d Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 0918fe84ee75 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 580fe375b210 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ ac7c0b5f316d Already exists                                                                                                                                                                                                                                           0.0s
   ✔ be859ab198fd Already exists                                                                                                                                                                                                                                           0.0s
   ✔ a4892c8f6a01 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ e1ea398d7112 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ bd1b279009de Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 19826778c030 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 64b77b0954e8 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 057e69972ce5 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 0b134091b624 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ ca38746399f3 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 165946a4077f Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 711b83a01973 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ c039e7ad6400 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 7771b8968c8f Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 2c597926e1df Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 9a5337321160 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ c0cae869da32 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ a7d099f012ad Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 91da5c4abc82 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 54d06b87c281 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 79694f5f894a Already exists                                                                                                                                                                                                                                           0.0s
   ✔ 78dc209bdd46 Already exists                                                                                                                                                                                                                                           0.0s
   ✔ b81782a80e0c Already exists                                                                                                                                                                                                                                           0.0s
[+] Creating 0/0
 ⠋ Service orthanc1  Building                                                                                                                                                                                                                                              0.1s
[+] Building 0.0s (10/10) FINISHED                                                                                                                                                                                                                         docker:desktop-linux
 => [orthanc1 internal] load build definition from orthanc-for-test.Dockerfile                                                                                                                                                                                             0.0s
 => => transferring dockerfile: 225B                                                                                                                                                                                                                                       0.0s
 => [orthanc1 internal] load metadata for docker.io/orthancteam/orthanc:24.5.1                                                                                                                                                                                             0.0s
 => [orthanc1 internal] load .dockerignore                                                                                                                                                                                                                                 0.0s
 => => transferring context: 79B                                                                                                                                                                                                                                           0.0s
 => [orthanc1 1/4] FROM docker.io/orthancteam/orthanc:24.5.1                                                                                                                                                                                                               0.0s
 => [orthanc1 internal] load build context                                                                                                                                                                                                                                 0.0s
 => => transferring context: 3.10kB                                                                                                                                                                                                                                        0.0s
 => CACHED [orthanc1 2/4] RUN pip install httpx pydicom  --break-system-packages                                                                                                                                                                                           0.0s
 => CACHED [orthanc1 3/4] COPY pyorthanc /python/pyorthanc                                                                                                                                                                                                                 0.0s
 => CACHED [orthanc1 4/4] COPY tests/orthanc/script.py /python/script.py                                                                                                                                                                                                   0.0s
 => [orthanc1] exporting to image                                                                                                                                                                                                                                          0.0s
 => => exporting layers                                                                                                                                                                                                                                                    0.0s
 => => writing image sha256:e604fd7dc35d1ff5ca788f0245783433e5d04622b615137ad70c9cf247e16c64                                                                                                                                                                               0.0s
[+] Creating 4/2 docker.io/library/pyorthanc-orthanc1                                                                                                                                                                                                                      0.0s
 ✔ Service orthanc1                Built                                                                                                                                                                                                                                   0.2s
[…]
