### 2025-01-11 gacou54/pyorthanc#74 @sinhaharsh
[JOSS] Please include a jupyter notebook based tutorial 
https://github.com/openjournals/joss-reviews/issues/7562

I acknowledge that there are multiple examples in the README and documentation.

However, it would be even better if there is a jupyter notebook or self-contained walkthrough to understand the usage of the software. If possible, please include examples of intended real-world usage of the software.
