### 2025-01-11 #7562 @sinhaharsh
Thank you @gacou54 @sappelhoff  I didn't have any trouble installing or testing the basic functionality for PyOrthanc. I have added a couple of suggestions for improvement.
