### 2025-01-13 gacou54/pyorthanc#73 @jrkerns
I also failed to run these unless I set `package-mode = false` in `pyproject.toml`.
