### 2025-01-13 d4b723a
Update tests with new functionalities from API REST 26
