+++ README.md
@@ -87,7 +87,8 @@ Note that recent PyOrthanc versions will likely support older Orthanc version.
 
 | PyOrthanc version | Generated from                                |
 |-------------------|-----------------------------------------------|
-| \>= 1.18.0        | Orthanc API 1.12.4 with Python Plugin 4.2     |
+| \>= 1.19.0        | Orthanc API 1.12.5 with Python Plugin 4.2     |
+| 1.18.0            | Orthanc API 1.12.4 with Python Plugin 4.2     |
 | 1.17.0            | Orthanc API 1.12.3 with Python Plugin 4.2     |
 | 1.13.2 to 1.16.1  | Orthanc API 1.12.1 with Python Plugin 4.1     |
 | 1.13.0, 1.13.1    | Orthanc API 1.12.1 with Python Plugin 4.0     |
