### 2025-01-13 gacou54/pyorthanc#73 @gacou54
Good catch! I had the same issue (my guess is that the new version of poetry break this).

I fix this in https://github.com/gacou54/pyorthanc/pull/71
