### 2025-01-13 gacou54/pyorthanc#74 @gacou54
I agree, I am improving the documentation in https://github.com/gacou54/pyorthanc/pull/72.

I'll keep you posted.
