### 2025-01-14 gacou54/pyorthanc#73 @gacou54
#71 has been merge in main and should fix this issue.
