### 2025-01-14 34dbb54
Add information on how tests are run
