### 2025-01-14 b826429
Add information on how tests are run
