### 2025-01-14 14176c6
✨ Add upload function, an easier way to upload DICOM data
