### 2025-01-14 ee9a545
✨ Add upload function, an easier way to upload DICOM data
