+++ pyorthanc/__init__.py
@@ -1,16 +1,20 @@
+# First imported to avoid circular import
+from .async_client import AsyncOrthanc
+from .client import Orthanc
+
 from . import errors, util
 from ._filtering import find, trim_patients
 from ._find import find_instances, find_patients, find_series, find_studies, query_orthanc
 from ._modality import Modality, RemoteModality
 from ._resources import Instance, Patient, Series, Study
-from .async_client import AsyncOrthanc
-from .client import Orthanc
+from ._upload import async_upload, upload
 from .jobs import Job
 from .retrieve import retrieve_and_write_instance, retrieve_and_write_patient, retrieve_and_write_patients, \
     retrieve_and_write_series, retrieve_and_write_study
 
 __all__ = [
     'AsyncOrthanc',
+    'async_upload',
     'Orthanc',
     'Modality',
     'RemoteModality',
@@ -31,6 +35,7 @@ __all__ = [
     'retrieve_and_write_study',
     'retrieve_and_write_series',
     'retrieve_and_write_instance',
+    'upload',
     'util',
     'errors',
 ]
