### 2025-01-15 70c4a2b
Add `Modality.find` alias and add modality notebook example
