+++ tests/test_modality.py
@@ -27,6 +27,26 @@ def test_echo(modality):
     assert result
 
 
+def test_find(modality):
+    setup_data(ORTHANC_2)
+    expected_query_answer = {
+        '0008,0005': {'Name': 'SpecificCharacterSet', 'Type': 'String', 'Value': 'ISO_IR 100'},
+        '0008,0050': {'Name': 'AccessionNumber', 'Type': 'String', 'Value': '20090926001'},
+        '0008,0052': {'Name': 'QueryRetrieveLevel', 'Type': 'String', 'Value': 'STUDY'},
+        '0008,0054': {'Name': 'RetrieveAETitle', 'Type': 'String', 'Value': 'ORTHANC'},
+        '0010,0020': {'Name': 'PatientID', 'Type': 'String', 'Value': 'MP15-067'},
+        '0020,000d': {'Name': 'StudyInstanceUID', 'Type': 'String', 'Value': '1.3.6.1.4.1.22213.2.6291.2.1'}
+    }
+
+    result = modality.find(PAYLOAD)
+
+    assert 'ID' in result.keys()
+    assert 'Path' in result.keys()
+
+    queries_answers = modality.get_query_answers()[result['ID']]
+
+    assert expected_query_answer == queries_answers
+
 def test_query(modality):
     setup_data(ORTHANC_2)
     expected_query_answer = {
@@ -56,6 +76,33 @@ def test_failed_query(modality):
 
 
 def test_move(modality):
+    expected_move_answer = {
+        'Description': 'REST API',
+        'LocalAet': ORTHANC_1.AeT,
+        'RemoteAet': ORTHANC_2.AeT,
+        'Query': [{
+            '0008,0050': '20090926001',
+            '0008,0052': 'STUDY',
+            '0010,0020': 'MP15-067',
+            '0020,000d': '1.3.6.1.4.1.22213.2.6291.2.1'
+        }],
+    }
+    setup_data(ORTHANC_2)
+
+    query_result = modality.query(PAYLOAD)
+    result = modality.move(query_result['ID'])
+
+    assert result == expected_move_answer
+
+    resulting_patient_information = modality.client.get_patients_id(
+        modality.client.get_patients()[0]
+    )
+
+    assert {k: v for k, v in PATIENT_INFORMATION.items() if k not in ['LastUpdate']} == \
+           {k: v for k, v in resulting_patient_information.items() if k not in ['LastUpdate']}
+
+
+def test_move_to_target_modality(modality):
     cmove_data = {'TargetAet': ORTHANC_1.AeT}
     expected_move_answer = {
         'Description': 'REST API',
