### 2025-01-15 f812a27
Remove old example
