### 2025-01-15 gacou54/pyorthanc#74 @gacou54
I added two jupyter notebook of the typical use-cases with Orthanc here https://github.com/gacou54/pyorthanc/tree/docs/examples.

- `find_data.ipynb` covers the most typical usage, which is parsing/finding data in Orthanc
- `modalities.ipynb` covers another common usage yet more "advanced" topic, which is when another DICOM server is connected to an Orthanc server

I will refer to them in the new updated documentation.

Does this fit what you had in mind?
