+++ README.md
@@ -169,6 +169,10 @@ This command starts 3 containers :
 * [Query (C-Find) and Retrieve (C-Move) from remote modality](docs/tutorial/quickstart.md#query-c-find-and-retrieve-c-move-from-remote-modality)
 ### [Advanced examples](docs/tutorial/advanced.md)
 ### [Releases](https://github.com/gacou54/pyorthanc/releases)
+### [Community guidelines](docs/contributing.md)
+* [Report an issue](docs/contributing.md#report-an-issue)
+* [Support](docs/contributing.md#seeking-support)
+* [Contribute](docs/contributing.md#contribute)
 ## [Contacts](docs/contacts.md#contacts)
 * [Maintainers Team](docs/contacts.md#maintainers-team)
 * [Useful links](docs/contacts.md#useful-links)
