+++ docs/tutorial/quickstart.md
@@ -1,10 +1,14 @@
 # First steps
 
 ## Orthanc installation
-For testing purposes, you can use
+For testing purposes, you can use one of this solution to install Orthanc
 - Orthanc's demo server (https://orthanc.uclouvain.be/demo)
-- Or deploy it locally with docker using `docker run -p 8042:8042 -p 4242:4242 orthancteam/orthanc`
-  - Other method to install Orthanc are available [here](https://www.orthanc-server.com/download.php)
+- Deploy it locally with docker using `docker run -p 8042:8042 -p 4242:4242 orthancteam/orthanc`
+- Using the docker-compose from the pyorthanc repo to ha
+  1. For this, clone the repo (`git clone https://github.com/gacou54/pyorthanc`)
+  2. Got in the repo (`cd pyorthanc`)
+  3. Start orthanc (`docker compose up orthanc1`)
+- Other method to install Orthanc are available [here](https://www.orthanc-server.com/download.php)
 
 ## PyOrthanc installation
 ```bash
