### 2025-01-16 8105228
Add precision regarding accronyms (CT, PACS, DICOMSE)
