### 2025-01-16 gacou54/pyorthanc#69 @gacou54
Fixed in new doc release #72
