### 2025-01-16 gacou54/pyorthanc#70 @gacou54
I added the link https://www.orthanc-server.com/download.php in the documentation in #72

It is now specified that the tests are ran in a docker container here https://github.com/gacou54/pyorthanc?tab=readme-ov-file#running-tests. Test are ran with Orthanc version ` orthancteam/orthanc:24.12.0` as specified in the docker-compose.yml file.
