+++ docs/cheat_sheet.md
@@ -2,220 +2,224 @@
 
 [Orthanc](https://www.orthanc-server.com/) has a very useful
 [cheat sheet](https://orthanc.uclouvain.be/book/users/rest-cheatsheet.html)
-for finding typical functionality and REST API routes. This page is an extension of that.
+for finding typical functionality and REST API routes. This page contains the link to the appropriate method in the `pyorthanc.Orthanc` class.
 
-| Path                                                 | GET                                                                   | POST                                                              | DELETE                                                   | PUT                                                   | Summary                                         |
-|:-----------------------------------------------------|:----------------------------------------------------------------------|:------------------------------------------------------------------|:---------------------------------------------------------|:------------------------------------------------------|:------------------------------------------------|
-| `/changes`                                           | `orthanc.get_changes()`                                               | -                                                                 | `orthanc.delete_changes()`                               | -                                                     | List changes                                    |
-| `/exports`                                           | `orthanc.get_exports()`                                               | -                                                                 | `orthanc.delete_exports()`                               | -                                                     | List exports                                    |
-| `/instances`                                         | `orthanc.get_instances()`                                             | `orthanc.post_instances()`                                        | -                                                        | -                                                     | List the available instances                    |
-| `/instances/{id}`                                    | `orthanc.get_instances_id(id)`                                        | -                                                                 | `orthanc.delete_instances_id(id)`                        | -                                                     | Get information about some instance             |
-| `/instances/{id}/anonymize`                          | -                                                                     | `orthanc.post_instances_id_anonymize(id)`                         | -                                                        | -                                                     | Anonymize instance                              |
-| `/instances/{id}/attachments`                        | `orthanc.get_instances_id_attachments(id)`                            | -                                                                 | -                                                        | -                                                     | List attachments                                |
-| `/instances/{id}/attachments/{name}`                 | `orthanc.get_instances_id_attachments_name(id, name)`                 | -                                                                 | `orthanc.delete_instances_id_attachments_name(id, name)` | `orthanc.put_instances_id_attachments_name(id, name)` | List operations on attachments                  |
-| `/instances/{id}/attachments/{name}/compress`        | -                                                                     | `orthanc.post_instances_id_attachments_name_compress(id, name)`   | -                                                        | -                                                     | Compress attachment                             |
-| `/instances/{id}/attachments/{name}/compressed-data` | `orthanc.get_instances_id_attachments_name_compressed_data(id, name)` | -                                                                 | -                                                        | -                                                     | Get attachment (no decompression)               |
-| `/instances/{id}/attachments/{name}/compressed-md5`  | `orthanc.get_instances_id_attachments_name_compressed_md5(id, name)`  | -                                                                 | -                                                        | -                                                     | Get MD5 of attachment on disk                   |
-| `/instances/{id}/attachments/{name}/compressed-size` | `orthanc.get_instances_id_attachments_name_compressed_size(id, name)` | -                                                                 | -                                                        | -                                                     | Get size of attachment on disk                  |
-| `/instances/{id}/attachments/{name}/data`            | `orthanc.get_instances_id_attachments_name_data(id, name)`            | -                                                                 | -                                                        | -                                                     | Get attachment                                  |
-| `/instances/{id}/attachments/{name}/info`            | `orthanc.get_instances_id_attachments_name_info(id, name)`            | -                                                                 | -                                                        | -                                                     | Get info about the attachment                   |
-| `/instances/{id}/attachments/{name}/is-compressed`   | `orthanc.get_instances_id_attachments_name_is_compressed(id, name)`   | -                                                                 | -                                                        | -                                                     | Is attachment compressed?                       |
[…]
