### 2025-01-17 7845d99
Add links too the more convenient methods
