### 2025-01-21 3a72bf8
🐛 Fix and Improve modality results
