### 2025-01-21 d291461
🔥 Add utility function to delete queries
