### 2025-01-21 1f96ca0
🔥 Add utility function to delete queries
