### 2025-01-21 gacou54/pyorthanc#74 @gacou54
I will close the issue considering the examples. If they don't match what you expected, feel free to reopen the issue.
