### 2025-01-22 498e036
Change `Modality` output for easier comprehension
