### 2025-02-03 #7562 @sappelhoff
@gacou54 please let me know once you finished your items from the list above: https://github.com/openjournals/joss-reviews/issues/7562#issuecomment-2632303964

I will then work on my final items, and then it's (hopefully) finally time to ship it!
