### 2025-02-15 #7562 @sappelhoff
Hi @gacou54 thanks for your patience!

- [ ] I see that you have fixed the version in the [Zenodo archive,](https://zenodo.org/records/14802226) thanks! Would you please also adjust the title of that archive to match the one of the paper? (see also the corresponding [checklist item above](https://github.com/openjournals/joss-reviews/issues/7562#issuecomment-2632303964))
