### 2025-02-15 #7562 @sappelhoff
@gacou54 -- I have gone through the remaining items and I am happy with all of them. As soon as https://github.com/openjournals/joss-reviews/issues/7562#issuecomment-2661097802 is addressed, I will recommend this for acceptance.
