### 2025-02-17 #7562 @editorialbot
:wave: @openjournals/bcm-eics, this paper is ready to be accepted and published.
